# app/public/swagger.json
{
  "consumes": [
    "application/json",
    "multipart/form-data"
  ],
  "produces": [
    "application/json"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Documentation of PetStore API.",
    "title": "PetStore.",
    "version": "1.0.0"
  },
  "basePath": "/",
  "paths": {
...
    "/pet/login": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Авторизация пользователя для работы с питомцами.",
        "operationId": "LoginRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/LoginRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/LoginResponse"
          }
        }
      }
    },
    "/pet/register": {
      "post": {
        "tags": [
          "auth"
        ],
        "summary": "Регистрация пользователя для работы с питомцами.",
        "operationId": "RegisterRequest",
        "parameters": [
          {
            "name": "Body",
            "in": "body",
            "schema": {
              "$ref": "#/definitions/RegisterRequest"
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/RegisterReponse"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "LoginRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "projects/LDmitryLD/petstore/app/internal/modules/auth/controller"
    },
    "LoginResponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "projects/LDmitryLD/petstore/app/internal/modules/auth/controller"
    },
...
    "RegisterReponse": {
      "type": "object",
      "properties": {
        "message": {
          "type": "string",
          "x-go-name": "Message"
        },
        "success": {
          "type": "boolean",
          "x-go-name": "Success"
        }
      },
      "x-go-package": "projects/LDmitryLD/petstore/app/internal/modules/auth/controller"
    },
    "RegisterRequest": {
      "type": "object",
      "properties": {
        "password": {
          "type": "string",
          "x-go-name": "Password"
        },
        "username": {
          "type": "string",
          "x-go-name": "Username"
        }
      },
      "x-go-package": "projects/LDmitryLD/petstore/app/internal/modules/auth/controller"
    },
...
  },
  "responses": {
...
    "LoginResponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/LoginResponse"
      }
    },
...
    "RegisterReponse": {
      "description": "",
      "schema": {
        "$ref": "#/definitions/RegisterReponse"
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  }
}